ink information refactor, split adminwidget out
