+++ packages/react-app/src/components/index.js
@@ -12,3 +12,4 @@ export { default as Wallet } from './Wallet.js'
 export { default as Blockie } from './Blockie.js'
 export { default as Header } from './Header.js'
 export { default as Timeline } from './Timeline.js'
+export { default as AdminWidget } from './AdminWidget.js'
